Add files via upload
For this project, I looked at Apple and Android app data in order to make recommendations for a hypothetical app developer interested in creating a new app. 
I explored this data using tableau, honing in on how much the app should cost, which category of app to create, and which platform to make the app available on.